@@ -26,7 +26,7 @@ module.exports = function (config)
             suppressAll: true // removes the duplicated traces
         },
         coverageReporter   : {
-            dir      : require('path').join(__dirname, './coverage/angular11'),
+            dir      : require('path').join(__dirname, './coverage/angular12'),
             subdir   : '.',
             reporters: [
                 {type: 'html'},
