@@ -1,5 +1,3 @@
-import { AuthOptions } from '@aws-amplify/auth/lib/types';
-
 export interface Environment {
     production?: boolean;
 
@@ -9,15 +7,14 @@ export interface Environment {
         apiUrl: string
     };
 
-    // AWS Cognito Configuration
-    amplify: {
-        Auth: AuthOptions
-    };
-
     // Contentful CMS
     contentful: {
         spaceId: string,
         token: string;
         environment?: string;  // defaults to `master`
+    };
+
+    push: {
+        publicKey: string
     };
 }
