@@ -7,7 +7,7 @@ import { Inject, Injectable } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
 import { ENV } from '@shared/constants';
 import { Environment } from '@shared/environment.model';
-import { GroupMemberSimple, GroupTypeRole, GroupWithChildren, NewGroupMemberForm } from '../group.model';
+import { GroupMemberSimple, GroupType, GroupTypeRole, GroupWithChildren, NewGroupMemberForm } from '../group.model';
 import { Observable } from 'rxjs';
 import { ApiResponse } from '@shared/shared.models';
 import { map } from 'rxjs/operators';
@@ -60,10 +60,11 @@ export class GroupsDataService extends HttpBaseService
         const personId = model.person.id;
         const groupId = model.groupId;
         const groupRoleId = model.groupRole;
+        const communicationPreference = model.communicationPreference;
         const firstVisitDate = model.firstVisitDate;
 
         const body = {
-            personId, groupId, groupRoleId, firstVisitDate
+            personId, groupId, groupRoleId, communicationPreference, firstVisitDate
         };
 
         return super.post<ApiResponse>(`${this._apiUrl}/v1/groups/${model.groupId}/add-member`, body)
@@ -75,5 +76,13 @@ export class GroupsDataService extends HttpBaseService
     registerAttendance$( model: GroupAttendanceForm ): Observable<any>
     {
         return super.post<any>(`${this._apiUrl}/v1/groups/${model.groupId}/attendance`, model);
+    }
+
+    getGroupType$(groupTypeId: number) : Observable<GroupType>
+    {
+        return super.get<ApiResponse>(`${this._apiUrl}/v1/groups/types/${groupTypeId}`, null)
+            .pipe(
+                map(response => response.data)
+            );
     }
 }
