@@ -8,9 +8,11 @@ import {
     SimpleChanges,
     ViewEncapsulation
 } from '@angular/core';
-import { GroupWithChildren } from '@features/admin/groups';
+import { Group, GroupWithChildren } from '@features/admin/groups';
 import { FlatTreeControl } from '@angular/cdk/tree';
 import { MatTreeFlatDataSource, MatTreeFlattener } from '@angular/material/tree';
+import { MatDialog } from '@angular/material/dialog';
+import { NewGroupDialogComponent } from '@features/admin/groups/manage/components/new/new-group-dialog.component';
 
 interface FlatNode {
     expandable: boolean;
@@ -29,6 +31,7 @@ export class GroupsViewerComponent implements OnChanges
 {
     @Input() groups: GroupWithChildren[] = [];
     @Output() selectedGroup = new EventEmitter<GroupWithChildren>();
+    @Output() addedGroup = new EventEmitter<Group>();
 
     treeControl = new FlatTreeControl<FlatNode>(node => node.level, node => node.expandable);
 
@@ -36,9 +39,62 @@ export class GroupsViewerComponent implements OnChanges
 
     /** Map from flat node to nested node. This helps us finding the nested node to be modified */
     flatNodeMap = new Map<number, GroupWithChildren>();
-
+    hasChild = (_: number, node: FlatNode) => node.expandable;
     private _treeFlattener: MatTreeFlattener<GroupWithChildren, FlatNode>;
 
+    /**
+     * Constructor
+     */
+    constructor(private _matDialog: MatDialog)
+    {
+        this._treeFlattener = new MatTreeFlattener(
+            this._transformer, node => node.level, node => node.expandable, node => node.groups);
+
+        this.dataSource = new MatTreeFlatDataSource(this.treeControl, this._treeFlattener);
+    }
+
+    // -----------------------------------------------------------------------------------------------------
+    // @ Lifecycle hooks
+    // -----------------------------------------------------------------------------------------------------
+
+    ngOnChanges( changes: SimpleChanges ): void
+    {
+        if ( changes['groups'] ) {
+            this.dataSource.data = changes['groups'].currentValue;
+        }
+    }
+
+    // -----------------------------------------------------------------------------------------------------
+    // @ Public methods
+    // -----------------------------------------------------------------------------------------------------
+
+    onSelectGroup( { item }: FlatNode ): void
+    {
+        const selectedGroup = this.flatNodeMap.get(item.id);
+        this.selectedGroup.emit(selectedGroup);
+    }
+
+    /**
+     * Open new group dialog
+     */
+    openAddGroupDialog({ item }: FlatNode): void
+    {
+        // Open the dialog
+        const dialogRef = this._matDialog.open(NewGroupDialogComponent, {
+            data : {
+                parentGroup: item
+            }
+        });
+
+        dialogRef.afterClosed()
+            .subscribe((result) => {
+                console.log('Compose dialog was closed!', result);
+            });
+    }
+
+    // -----------------------------------------------------------------------------------------------------
+    // @ Private methods
+    // -----------------------------------------------------------------------------------------------------
     private _transformer = (node: GroupWithChildren, level: number) => {
         // While transforming add the item to the map
         this.flatNodeMap.set(node.id, node);
@@ -50,28 +106,5 @@ export class GroupsViewerComponent implements OnChanges
             // optional
             item: node
         };
-    }
-
-    hasChild = (_: number, node: FlatNode) => node.expandable;
-
-    constructor()
-    {
-        this._treeFlattener = new MatTreeFlattener(
-            this._transformer, node => node.level, node => node.expandable, node => node.groups);
-
-        this.dataSource = new MatTreeFlatDataSource(this.treeControl, this._treeFlattener);
-    }
-
-    ngOnChanges( changes: SimpleChanges ): void
-    {
-        if ( changes['groups'] ) {
-            this.dataSource.data = changes['groups'].currentValue;
-        }
-    }
-
-    onSelectGroup( { item }: FlatNode ): void
-    {
-        const selectedGroup = this.flatNodeMap.get(item.id);
-        this.selectedGroup.emit(selectedGroup);
     }
 }
