@@ -1,18 +1,21 @@
 import {
     ChangeDetectionStrategy,
-    Component, EventEmitter,
+    Component,
+    EventEmitter,
     Input,
     OnChanges,
     Output,
     SimpleChanges,
+    ViewChild,
     ViewEncapsulation
 } from '@angular/core';
 import { GroupMemberSimple, GroupMembersSimple, GroupWithChildren, NewGroupMemberForm } from '@features/admin/groups';
 import { MatTableDataSource } from '@angular/material/table';
 import { SelectionModel } from '@angular/cdk/collections';
-import { FormControl } from '@angular/forms';
 import { MatDialog } from '@angular/material/dialog';
 import { AddGroupMemberFormDialogComponent } from './add/add-group-member-form-dialog.component';
+import { MatSort } from '@angular/material/sort';
+import { MatPaginator } from '@angular/material/paginator';
 
 @Component({
     selector       : 'group-members',
@@ -20,22 +23,32 @@ import { AddGroupMemberFormDialogComponent } from './add/add-group-member-form-d
     encapsulation  : ViewEncapsulation.None,
     changeDetection: ChangeDetectionStrategy.OnPush
 })
-export class GroupMembersComponent implements  OnChanges
+export class GroupMembersComponent implements OnChanges
 {
     @Input() members: GroupMembersSimple;
     @Input() group: GroupWithChildren;
     @Input() isLoading: boolean = false;
     @Output() memberAdded = new EventEmitter<NewGroupMemberForm>();
 
+    // Fix: https://stackoverflow.com/questions/46786757/angular-matsort-does-not-sort
+    private _sort: MatSort;
+    @ViewChild(MatSort) set matSort(ms: MatSort) {
+        this._sort = ms;
+        this._setDataSourceAttributes();
+    }
+    private _paginator: MatPaginator
+    @ViewChild(MatPaginator) set matPaginator(mp: MatPaginator) {
+        this._paginator = mp;
+        this._setDataSourceAttributes();
+    }
+
     dialogRef: any;
 
-    displayedColumns: string[] = ['select', 'photoUrl', 'firstName', 'lastName', 'gender', 'groupMemberRole'];
+    displayedColumns: string[] = ['select', 'photoUrl', 'firstName', 'lastName', 'gender', 'groupMemberRole', 'recordStatus'];
 
     dataSource: MatTableDataSource<GroupMemberSimple> = new MatTableDataSource([]);
 
     selection = new SelectionModel<GroupMemberSimple>(true, []);
-
-    searchInputControl: FormControl = new FormControl();
 
     constructor(private _matDialog: MatDialog) {
     }
@@ -86,5 +99,16 @@ export class GroupMembersComponent implements  OnChanges
                 // Do something here
                 this.memberAdded.emit(formData);
             });
+    }
+
+    applyFilter(event: Event) {
+        const filterValue = (event.target as HTMLInputElement).value;
+        this.dataSource.filter = filterValue.trim().toLowerCase();
+    }
+
+    private _setDataSourceAttributes() : void
+    {
+        this.dataSource.paginator = this._paginator;
+        this.dataSource.sort = this._sort;
     }
 }
