@@ -1,6 +1,6 @@
-/* tslint:disable:max-line-length */
+/* eslint-disable */
 
-// 20210326 - 1777 icons
+// Updated at: 20210425 - 1792 icons
 export const material = [
     '10k',
     '10mp',
@@ -100,6 +100,7 @@ export const material = [
     'addchart',
     'adjust',
     'admin_panel_settings',
+    'ads_click',
     'agriculture',
     'air',
     'airline_seat_flat',
@@ -146,6 +147,7 @@ export const material = [
     'apps',
     'architecture',
     'archive',
+    'area_chart',
     'arrow_back',
     'arrow_back_ios',
     'arrow_back_ios_new',
@@ -195,6 +197,7 @@ export const material = [
     'autorenew',
     'av_timer',
     'baby_changing_station',
+    'back_hand',
     'backpack',
     'backspace',
     'backup',
@@ -386,6 +389,7 @@ export const material = [
     'compare',
     'compare_arrows',
     'compass_calibration',
+    'compost',
     'compress',
     'computer',
     'confirmation_number',
@@ -429,10 +433,12 @@ export const material = [
     'crop_portrait',
     'crop_rotate',
     'crop_square',
+    'cruelty_free',
     'dangerous',
     'dark_mode',
     'dashboard',
     'dashboard_customize',
+    'data_exploration',
     'data_saver_off',
     'data_saver_on',
     'data_usage',
@@ -482,6 +488,7 @@ export const material = [
     'directions_walk',
     'dirty_lens',
     'disabled_by_default',
+    'disabled_visible',
     'disc_full',
     'dns',
     'do_disturb',
@@ -521,6 +528,7 @@ export const material = [
     'draw',
     'drive_eta',
     'drive_file_move',
+    'drive_file_move_rtl',
     'drive_file_rename_outline',
     'drive_folder_upload',
     'dry',
@@ -533,7 +541,6 @@ export const material = [
     'earbuds',
     'earbuds_battery',
     'east',
-    'eco',
     'edgesensor_high',
     'edgesensor_low',
     'edit',
@@ -715,6 +722,7 @@ export const material = [
     'foundation',
     'free_breakfast',
     'free_cancellation',
+    'front_hand',
     'fullscreen',
     'fullscreen_exit',
     'functions',
@@ -829,6 +837,7 @@ export const material = [
     'import_export',
     'important_devices',
     'inbox',
+    'incomplete_circle',
     'indeterminate_check_box',
     'info',
     'input',
@@ -1215,6 +1224,8 @@ export const material = [
     'pie_chart_outline',
     'pin',
     'pin_drop',
+    'pin_end',
+    'pin_invoke',
     'pivot_table_chart',
     'place',
     'plagiarism',
@@ -1289,6 +1300,7 @@ export const material = [
     'recent_actors',
     'recommend',
     'record_voice_over',
+    'recycling',
     'redeem',
     'redo',
     'reduce_capacity',
@@ -1631,6 +1643,7 @@ export const material = [
     'timer_3',
     'timer_3_select',
     'timer_off',
+    'tips_and_updates',
     'title',
     'toc',
     'today',
@@ -1737,8 +1750,10 @@ export const material = [
     'watch_later',
     'water',
     'water_damage',
+    'water_drop',
     'waterfall_chart',
     'waves',
+    'waving_hand',
     'wb_auto',
     'wb_cloudy',
     'wb_incandescent',
@@ -4152,87 +4167,91 @@ export const feather = [
     'zoom-in',
     'zoom-out'
 ];
-// heroicons v0.4.2
+// heroicons v1.0.1 - 230 icons
 export const heroicons = [
     'academic-cap',
     'annotation',
-    'arrow-circle-down',
     'adjustments',
+    'archive',
+    'arrow-circle-down',
     'arrow-circle-left',
     'arrow-circle-right',
-    'archive',
     'arrow-circle-up',
     'arrow-down',
-    'arrow-narrow-down',
     'arrow-left',
+    'arrow-narrow-down',
     'arrow-narrow-left',
     'arrow-narrow-right',
-    'arrow-right',
     'arrow-narrow-up',
+    'arrow-right',
+    'arrow-sm-left',
+    'arrow-sm-down',
+    'arrow-sm-right',
+    'arrow-sm-up',
     'arrow-up',
-    'at-symbol',
     'arrows-expand',
-    'backspace',
+    'at-symbol',
     'badge-check',
+    'backspace',
     'ban',
     'beaker',
     'bell',
     'book-open',
     'bookmark-alt',
     'bookmark',
-    'calendar',
     'briefcase',
-    'calculator',
     'cake',
+    'calculator',
+    'calendar',
     'camera',
     'cash',
     'chart-bar',
     'chart-pie',
     'chart-square-bar',
     'chat-alt-2',
+    'chat-alt',
     'chat',
     'check-circle',
-    'chat-alt',
     'check',
     'chevron-double-down',
     'chevron-double-left',
-    'chevron-down',
     'chevron-double-up',
-    'chevron-left',
     'chevron-double-right',
+    'chevron-down',
+    'chevron-left',
     'chevron-right',
     'chevron-up',
     'chip',
     'clipboard-check',
     'clipboard-copy',
     'clipboard-list',
-    'clock',
-    'cloud-upload',
     'clipboard',
+    'clock',
     'cloud-download',
-    'code',
+    'cloud-upload',
     'cloud',
+    'code',
     'cog',
     'collection',
-    'color-swatch',
     'credit-card',
+    'color-swatch',
     'cube-transparent',
     'cube',
     'currency-bangladeshi',
     'currency-dollar',
-    'currency-pound',
     'currency-euro',
+    'currency-pound',
     'currency-rupee',
-    'cursor-click',
     'currency-yen',
+    'cursor-click',
     'database',
     'desktop-computer',
     'device-mobile',
     'device-tablet',
-    'document-download',
     'document-add',
-    'document-remove',
+    'document-download',
     'document-duplicate',
+    'document-remove',
     'document-report',
     'document-search',
     'document-text',
@@ -4241,81 +4260,81 @@ export const heroicons = [
     'dots-horizontal',
     'dots-vertical',
     'download',
+    'duplicate',
     'emoji-happy',
     'emoji-sad',
-    'duplicate',
     'exclamation-circle',
     'exclamation',
     'external-link',
     'eye-off',
-    'film',
-    'fast-forward',
-    'finger-print',
     'eye',
+    'fast-forward',
+    'film',
     'filter',
+    'finger-print',
     'fire',
     'flag',
     'folder-add',
-    'folder-remove',
-    'folder-open',
-    'folder',
     'folder-download',
+    'folder-open',
+    'folder-remove',
+    'folder',
     'gift',
     'globe-alt',
-    'hand',
     'globe',
+    'hand',
     'hashtag',
-    'identification',
     'heart',
     'home',
+    'identification',
     'inbox-in',
     'inbox',
     'information-circle',
     'key',
+    'library',
     'light-bulb',
     'lightning-bolt',
-    'library',
     'link',
     'location-marker',
     'lock-closed',
     'lock-open',
-    'logout',
-    'mail',
     'login',
-    'map',
+    'logout',
     'mail-open',
+    'mail',
+    'map',
     'menu-alt-1',
+    'menu-alt-2',
     'menu-alt-3',
     'menu-alt-4',
-    'menu-alt-2',
     'menu',
     'microphone',
     'minus-circle',
     'minus-sm',
     'minus',
-    'music-note',
     'moon',
+    'music-note',
     'newspaper',
     'office-building',
-    'paper-clip',
     'paper-airplane',
+    'paper-clip',
     'pause',
     'pencil-alt',
     'pencil',
     'phone-incoming',
     'phone-missed-call',
     'phone-outgoing',
-    'plus-circle',
-    'play',
     'photograph',
     'phone',
+    'play',
+    'plus-circle',
     'plus-sm',
     'plus',
     'presentation-chart-bar',
-    'puzzle',
     'presentation-chart-line',
-    'qrcode',
     'printer',
+    'qrcode',
+    'puzzle',
     'question-mark-circle',
     'receipt-refund',
     'receipt-tax',
@@ -4327,8 +4346,8 @@ export const heroicons = [
     'save',
     'scale',
     'scissors',
-    'search',
     'search-circle',
+    'search',
     'selector',
     'server',
     'share',
@@ -4338,22 +4357,22 @@ export const heroicons = [
     'shopping-cart',
     'sort-ascending',
     'sort-descending',
-    'sparkles',
     'speakerphone',
+    'sparkles',
     'star',
-    'status-offline',
     'status-online',
+    'status-offline',
     'stop',
     'sun',
     'support',
-    'switch-horizontal',
     'switch-vertical',
+    'switch-horizontal',
     'table',
     'tag',
     'template',
-    'thumb-down',
     'terminal',
     'thumb-up',
+    'thumb-down',
     'ticket',
     'translate',
     'trash',
@@ -4362,22 +4381,22 @@ export const heroicons = [
     'truck',
     'upload',
     'user-add',
-    'user-circle',
     'user-group',
-    'user',
+    'user-circle',
     'user-remove',
+    'user',
     'users',
+    'variable',
     'video-camera',
     'view-boards',
-    'variable',
     'view-grid-add',
     'view-grid',
     'view-list',
-    'volume-off',
     'volume-up',
-    'wifi',
+    'volume-off',
     'x-circle',
-    'x',
     'zoom-in',
-    'zoom-out'
+    'wifi',
+    'zoom-out',
+    'x'
 ];
